+++ servicex/resources/internal/add_file_to_dataset.py
